+++ examples/flow_input.fs
@@ -1,0 +1,3 @@
+print "Press space"
+if input.is_key_down("SPACE"):
+    print "space pressed"