+++ examples/lua_timer.lua
@@ -1,0 +1,9 @@
+time.boot()
+time.mark()
+
+on_update(function()
+  if time.get_time() > 0.1 then
+    log("tick")
+    time.mark()
+  end
+end)